explode BodyValidator in multipart/unipart traits
