+++ src/PSR7/ResponseValidator.php
@@ -6,7 +6,7 @@ namespace OpenAPIValidation\PSR7;
 
 use cebe\openapi\spec\OpenApi;
 use OpenAPIValidation\PSR7\Exception\ValidationFailed;
-use OpenAPIValidation\PSR7\Validators\BodyValidator;
+use OpenAPIValidation\PSR7\Validators\BodyValidator\BodyValidator;
 use OpenAPIValidation\PSR7\Validators\HeadersValidator;
 use OpenAPIValidation\PSR7\Validators\ValidatorChain;
 use Psr\Http\Message\ResponseInterface;
