+++ src/PSR7/ServerRequestValidator.php
@@ -8,7 +8,7 @@ use cebe\openapi\spec\OpenApi;
 use OpenAPIValidation\PSR7\Exception\MultipleOperationsMismatchForRequest;
 use OpenAPIValidation\PSR7\Exception\NoOperation;
 use OpenAPIValidation\PSR7\Exception\ValidationFailed;
-use OpenAPIValidation\PSR7\Validators\BodyValidator;
+use OpenAPIValidation\PSR7\Validators\BodyValidator\BodyValidator;
 use OpenAPIValidation\PSR7\Validators\CookiesValidator;
 use OpenAPIValidation\PSR7\Validators\HeadersValidator;
 use OpenAPIValidation\PSR7\Validators\PathValidator;
